New file utils.js, moved teardown function
